Relay client Javascript and API python/request-server/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
Markdown==3.2.2
mysql-connector-python==8.0.20
mysqlclient==1.4.6
protobuf==3.12.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
